WIP: moving scenarios
