+++ app_test_suite/steps/scenarios/simple.py
@@ -1,195 +1,39 @@
 import argparse
 import logging
 import os
-import shutil
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from tempfile import TemporaryDirectory
-from typing import Set, Optional, List, cast
+from typing import Optional, Set, cast
 
 import configargparse
 import yaml
-from pykube import KubeConfig, HTTPClient, ConfigMap
+from pykube import HTTPClient, KubeConfig, ConfigMap
 from pytest_helm_charts.giantswarm_app_platform.custom_resources import AppCR
 from pytest_helm_charts.giantswarm_app_platform.entities import ConfiguredApp
 from pytest_helm_charts.giantswarm_app_platform.utils import (
-    delete_app,
-    wait_for_app_to_be_deleted,
     create_app,
     wait_for_apps_to_run,
+    delete_app,
+    wait_for_app_to_be_deleted,
 )
-from step_exec_lib.errors import ConfigError, ValidationError
-from step_exec_lib.steps import BuildStepsFilteringPipeline, BuildStep
-from step_exec_lib.types import Context, StepType, STEP_ALL
+from step_exec_lib.errors import ConfigError
+from step_exec_lib.steps import BuildStep
+from step_exec_lib.types import StepType, STEP_ALL, Context
 from step_exec_lib.utils.config import get_config_value_by_cmd_line_option
 from step_exec_lib.utils.processes import run_and_log
 
-from app_test_suite.cluster_providers.cluster_provider import ClusterInfo, ClusterType
-from app_test_suite.errors import ATSTestError
-from app_test_suite.steps.repositories import ChartMuseumAppRepository
-from app_test_suite.steps.test_types import config_option_cluster_type_for_test_type
 from cluster_manager import ClusterManager
-from steps.test_types import STEP_TEST_FUNCTIONAL, STEP_TEST_SMOKE
+from cluster_providers.cluster_provider import ClusterType, ClusterInfo
+from errors import ATSTestError
+from steps.base import TestExecutor, BaseTestScenariosFilteringPipeline, TestExecInfo, CONTEXT_KEY_CHART_YAML
+from steps.repositories import ChartMuseumAppRepository
+from steps.test_types import config_option_cluster_type_for_test_type, STEP_TEST_FUNCTIONAL, STEP_TEST_SMOKE
 
 TEST_APP_CATALOG_NAME: str = "chartmuseum"
+CONTEXT_KEY_APP_CR: str = "app_cr"
+CONTEXT_KEY_APP_CM_CR: str = "app_cm_cr"
+CHART_YAML = "Chart.yaml"
 
-context_key_chart_yaml: str = "chart_yaml"
-context_key_app_cr: str = "app_cr"
-context_key_app_cm_cr: str = "app_cm_cr"
-
-_chart_yaml = "Chart.yaml"
 logger = logging.getLogger(__name__)
-
-
-class BaseTestScenariosFilteringPipeline(BuildStepsFilteringPipeline):
-    """
-    Pipeline that combines all the steps required to run application tests.
-    """
-
-    key_config_group_name = "Base app testing options"
-    key_config_option_skip_deploy_app = "--app-tests-skip-app-deploy"
-    key_config_option_deploy_namespace = "--app-tests-deploy-namespace"
-    key_config_option_deploy_config_file = "--app-tests-app-config-file"
-
-    def __init__(self, pipeline: List[BuildStep], cluster_manager: ClusterManager):
-        super().__init__(pipeline, self.key_config_group_name)
-        self._cluster_manager = cluster_manager
-
-    def initialize_config(self, config_parser: configargparse.ArgParser) -> None:
-        super().initialize_config(config_parser)
-        config_parser.add_argument(
-            "-c",
-            "--chart-file",
-            required=True,
-            help="Path to the Helm Chart tar.gz file to test.",
-        )
-        if self._config_parser_group is None:
-            raise ValueError("'_config_parser_group' can't be None")
-        self._config_parser_group.add_argument(
-            self.key_config_option_skip_deploy_app,
-            required=False,
-            action="store_true",
-            help="Skip automated app deployment for the test run to the test cluster (using an App CR).",
-        )
-        self._config_parser_group.add_argument(
-            self.key_config_option_deploy_namespace,
-            required=False,
-            default="default",
-            help="The namespace your app under test should be deployed to for running tests.",
-        )
-        self._config_parser_group.add_argument(
-            self.key_config_option_deploy_config_file,
-            required=False,
-            help="Path for a configuration file (values file) for your app when it's deployed for testing.",
-        )
-        self._cluster_manager.initialize_config(self._config_parser_group)
-
-    def pre_run(self, config: argparse.Namespace) -> None:
-        super().pre_run(config)
-        if self._all_pre_runs_skipped:
-            return
-
-        if not config.chart_file or not os.path.isfile(config.chart_file):
-            raise ConfigError("chart-file", f"The file '{config.chart_file}' can't be found.")
-
-        self._cluster_manager.pre_run(config)
-        app_config_file = get_config_value_by_cmd_line_option(config, self.key_config_option_deploy_config_file)
-        if app_config_file:
-            if not os.path.isfile(app_config_file):
-                raise ATSTestError(
-                    f"Application test run was configured to use '{app_config_file}' as app"
-                    f" config file, but it doesn't exist."
-                )
-            try:
-                with open(app_config_file, "r") as file:
-                    yaml.safe_load(file)
-            except Exception:
-                raise ATSTestError(
-                    f"Application config file '{app_config_file}' found, but can't be loaded"
-                    f"as a correct YAML document."
-                )
-
-    def cleanup(
-        self,
-        config: argparse.Namespace,
-        context: Context,
-        has_build_failed: bool,
-    ) -> None:
-        self._cluster_manager.cleanup()
-
-
-@dataclass
-class TestExecInfo:
-    """
-    TestExecInfo provides all the information that is passed from a test scenario to test executor.
-    """
-
-    chart_path: str
-    """Path to the chart file under test."""
-    chart_ver: str
-    """Chart version detected from the chart."""
-    app_config_file_path: Optional[str]
-    """Path to an optional Helm values file used to configure the chart under test."""
-    cluster_type: str
-    """A string representing a cluster type that the test scenario is running on."""
-    cluster_version: str
-    """Kubernetes cluster version of a cluster the test scenario is running on."""
-    kube_config_path: str
-    """Path to kube.config to connect to the cluster."""
-    test_type: str
-    """Type of test to execute by the test executor."""
-    test_dir: str
-    """Path to a directory where test code is available."""
-
-
-class TestExecutor(ABC):
-    """
-    Base abstract class to implement different test executors.
-
-    Test executors are responsible for running actual tests in a scenario using a specific
-    test platform like `pytest` or `go test`.
-    """
-
-    def validate(self, config: argparse.Namespace, module_name: str) -> None:
-        """Validate any configuration related to the test executor."""
-        raise NotImplementedError()
-
-    def prepare_test_environment(self, exec_info: TestExecInfo) -> None:
-        """Optional step to prepare environment where your tests are executed (ie. installing dependencies)."""
-        raise NotImplementedError()
-
-    def execute_test(self, exec_info: TestExecInfo) -> None:
-        """Execute test using a specific test executor and information provided as exec_info."""
-        raise NotImplementedError()
-
-
-class TestInfoProvider(BuildStep):
-    """
-    Since the whole build pipeline can change Chart.yaml file multiple times, this
-    class loads the Chart.yaml as dict into context at the beginning of testing
-    pipeline.
-    """
-
-    @property
-    def steps_provided(self) -> Set[StepType]:
-        return {STEP_ALL}
-
-    def run(self, config: argparse.Namespace, context: Context) -> None:
-        with TemporaryDirectory(prefix="ats-") as tmp_dir:
-            shutil.unpack_archive(config.chart_file, tmp_dir)
-            _, sub_dirs, _ = next(os.walk(tmp_dir))
-            for sub_dir in sub_dirs:
-                chart_yaml_path = os.path.join(tmp_dir, sub_dir, "Chart.yaml")
-                if os.path.isfile(chart_yaml_path):
-                    with open(chart_yaml_path, "r") as file:
-                        chart_yaml = yaml.safe_load(file)
-                        logger.debug(f"Loading 'Chart.yaml' from subdirectory '{sub_dir}' in the chart archive.")
-                        context[context_key_chart_yaml] = chart_yaml
-                    break
-            else:
-                raise ValidationError(
-                    self.name, "Couldn't find 'Chart.yaml' in any subdirectory of the chart archive file."
-                )
 
 
 class SimpleTestScenario(BuildStep, ABC):
@@ -249,12 +93,12 @@ class SimpleTestScenario(BuildStep, ABC):
 
     def run_tests(self, config: argparse.Namespace, context: Context) -> None:
         app_config_file_path = get_config_value_by_cmd_line_option(
-            config, BaseTestScenariosFilteringPipeline.key_config_option_deploy_config_file
+            config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_DEPLOY_CONFIG_FILE
         )
         cluster_info = cast(ClusterInfo, self._cluster_info)
         exec_info = TestExecInfo(
             chart_path=config.chart_file,
-            chart_ver=context[context_key_chart_yaml]["version"],
+            chart_ver=context[CONTEXT_KEY_CHART_YAML]["version"],
             app_config_file_path=app_config_file_path,
             cluster_type=self._test_cluster_type,
             cluster_version=cluster_info.version,
@@ -361,7 +205,7 @@ class SimpleTestScenario(BuildStep, ABC):
         try:
             if (
                 not get_config_value_by_cmd_line_option(
-                    config, BaseTestScenariosFilteringPipeline.key_config_option_skip_deploy_app
+                    config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_SKIP_DEPLOY_APP
                 )
                 and not self._skip_app_deploy
             ):
@@ -371,25 +215,25 @@ class SimpleTestScenario(BuildStep, ABC):
             raise ATSTestError(f"Application deployment failed: {e}")
         finally:
             if not get_config_value_by_cmd_line_option(
-                config, BaseTestScenariosFilteringPipeline.key_config_option_skip_deploy_app
+                config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_SKIP_DEPLOY_APP
             ):
                 self._delete_app(config, context)
 
     def _deploy_tested_chart_as_app(self, config: argparse.Namespace, context: Context) -> None:
-        app_name = context[context_key_chart_yaml]["name"]
-        app_version = context[context_key_chart_yaml]["version"]
+        app_name = context[CONTEXT_KEY_CHART_YAML]["name"]
+        app_version = context[CONTEXT_KEY_CHART_YAML]["version"]
         deploy_namespace = get_config_value_by_cmd_line_option(
-            config, BaseTestScenariosFilteringPipeline.key_config_option_deploy_namespace
+            config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_DEPLOY_NAMESPACE
         )
         app_config_file_path = get_config_value_by_cmd_line_option(
-            config, BaseTestScenariosFilteringPipeline.key_config_option_deploy_config_file
+            config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_DEPLOY_CONFIG_FILE
         )
 
         app_obj = self._deploy_chart(
             app_name, app_version, deploy_namespace, app_config_file_path, TEST_APP_CATALOG_NAME
         )
-        context[context_key_app_cr] = app_obj.app
-        context[context_key_app_cm_cr] = app_obj.app_cm
+        context[CONTEXT_KEY_APP_CR] = app_obj.app
+        context[CONTEXT_KEY_APP_CM_CR] = app_obj.app_cm
 
     def _deploy_chart(
         self, app_name: str, app_version: str, deploy_namespace: str, app_config_file_path: str, app_catalog_name: str
@@ -423,14 +267,14 @@ class SimpleTestScenario(BuildStep, ABC):
     # noinspection PyMethodMayBeStatic
     def _delete_app(self, config: argparse.Namespace, context: Context) -> None:
         if get_config_value_by_cmd_line_option(
-            config, BaseTestScenariosFilteringPipeline.key_config_option_skip_deploy_app
+            config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_SKIP_DEPLOY_APP
         ):
             return
 
-        app_obj = cast(AppCR, context[context_key_app_cr])
+        app_obj = cast(AppCR, context[CONTEXT_KEY_APP_CR])
         values_cm = None
-        if context_key_app_cm_cr in context:
-            values_cm = cast(ConfigMap, context[context_key_app_cm_cr])
+        if CONTEXT_KEY_APP_CM_CR in context:
+            values_cm = cast(ConfigMap, context[CONTEXT_KEY_APP_CM_CR])
         logger.info("Deleting App CR and its values CM")
         delete_app(ConfiguredApp(app_obj, values_cm))
         wait_for_app_to_be_deleted(self._kube_client, app_obj.name, app_obj.namespace, self._APP_DELETION_TIMEOUT_SEC)
