+++ tests/pytest/test_pytest_runner.py
@@ -4,9 +4,9 @@ from typing import cast
 from pytest_mock import MockerFixture
 
 import app_test_suite
-from app_test_suite.steps.base_test_runner import context_key_chart_yaml, TEST_APP_CATALOG_NAME
-from app_test_suite.steps.pytest.pytest import UpgradeTestScenario
-from steps.base_test_runner import SmokeTestScenario
+from app_test_suite.steps.base import CONTEXT_KEY_CHART_YAML, TEST_APP_CATALOG_NAME
+from steps.scenarios.upgrade import UpgradeTestScenario
+from steps.scenarios.simple import SmokeTestScenario
 from app_test_suite.steps.upgrade_test_runner import STABLE_APP_CATALOG_NAME, KEY_PRE_UPGRADE, KEY_POST_UPGRADE
 from step_exec_lib.types import StepType
 from tests.helpers import (
@@ -46,7 +46,7 @@ def test_upgrade_pytest_runner_run(mocker: MockerFixture) -> None:
     config = get_base_config(mocker)
     configure_for_upgrade_test(config)
 
-    context = {context_key_chart_yaml: {"name": MOCK_APP_NAME, "version": MOCK_APP_VERSION}}
+    context = {CONTEXT_KEY_CHART_YAML: {"name": MOCK_APP_NAME, "version": MOCK_APP_VERSION}}
     runner = UpgradeTestScenario(mock_cluster_manager)
     runner.run(config, context)
 
@@ -91,7 +91,7 @@ def test_pytest_smoke_runner_run(mocker: MockerFixture) -> None:
     patch_pytest_test_runner(mocker, run_and_log_call_result_mock)
 
     config = get_base_config(mocker)
-    context = {context_key_chart_yaml: {"name": MOCK_APP_NAME, "version": MOCK_APP_VERSION}}
+    context = {CONTEXT_KEY_CHART_YAML: {"name": MOCK_APP_NAME, "version": MOCK_APP_VERSION}}
     runner = SmokeTestScenario(mock_cluster_manager)
     runner.run(config, context)
